feat(swift): extract node resolution from SurfaceViewModel into NodeResolver
Extract component tree resolution, schema property classification, check validation,
action binding, and dynamic value evaluation from SurfaceViewModel into a dedicated
NodeResolver class.

- Introduce NodeResolver under swift/core/Sources/A2UICore/Resolution/
- Update SurfaceViewModel to compose NodeResolver and publish rootNode
- Add unit tests for NodeResolver in NodeResolverTests

Fixes #2090
